Use correct URI to get Repository
In a server process we got a repositoryDir from the URI which not
contains an auth data. In Fetcher we should do the same, because an
uri with an auth data can be different from original uri (e.g. it can
contain a user name).

If we try to use the same bare repository for different URIes - we get
an exception from GitServerUtil.getRepository().