Use long-lived revisions cache to improve mapFullPath performance.
Entries in old cache lived only 10 seconds and we did a lot of commit
lookups. Most of these lookups went through all repository history and
didn't find a commit. Such lookups took a lot of time.

New cache isn't cleaned on timeout. We clean cache for a single
repository when new revision in this repository found.