Speedup getCurrentVersion() when fetch is in the server process
Previous implementation checked if remote ref is updated and if it is
- did the fetch. It was done this way because we didn't want to fire
separate process for fetch when fetch isn't required. It turns out
that establishing connection to remote repository through ssh has
significant cost and with previous implementation we establish
connection twice: once for listing remote refs and once for fetch.
Now we do fetch without listing remote refs if the fetch is made in
the server process.