TW-20953 Set maximum streamFileThreshold for fetch in separate process
This seems to be related to TW-14947. JGit streams files larger than
the specified threshold (50Mb by default) and it is very slow. We
introduced an internal parameter teamcity.git.stream.file.threshold.mb
for changing the threshold, but it had no effect on fetch in a
separate process. Now we set maximum threshold for the fetch in
separate process without an option, because it is better to get an
OutOfMemoryError and increase the memory for fetch, than to wait for
hours for fetch to finish.