TW-22369 support feature branches with agent-side checkout
If build parameters contain a branch name for VCS root, use it instead
of default VCS root branch. Server can give a hint on which branch to
use on agent using such parameters. If required revision is not found
in specified (or default) branch - fetch all branches and check again.