Optimize mapFullPath using hint commit
An IDE sends a merge-base commit between local changes on dev machine
and the remote branch in the fullPath, and git-plugin returns
non-empty paths for repositories which have such a commit.

A commit lookup is slow so we cache info on presence of the commit in
repository (with LRU cache). Problem is that each developer has its
own merge-base commit, so cache works only when search for suitable
configurations is repeated twice with the same merge-base.

To fix that IDE can also send a hint commit which is the same among
all developer machines. The only requirement for the hint commit is
that if repository doesn't contain it => it doesn't contain a
merge-base commit, i.e. hint commit should be some ancestor of the
merge-base. And it should be the same on most of dev machines, so info
on its presence won't be discarded from the cache and we will avoid
the most expensive operation: commit lookup in repository which
doesn't contain it.