Don't use -ff for 'git clean' on agent (revert 5c411cf)
When a build configuration has several VCS roots, one of them is
checked out to the root of the working directory and others are
checked out into subfolders, then clean in the first VCS root will
delete all files of nested VCS roots. Depending on an order of VCS
root handling, this leads to either clone from scratch in nested VCS
roots or a build failure due of lack of files in nested roots.