Fix patch in separate process for configurations with several VCS roots
GitPatchBuilderDispatcher wrote fromRevision even if it was null, it was
serialized as 'null', patch process was trying to find such a revision
and failed. GitPatchBuilder had a logic: if from revision is not found -
delete the directory where the VCS root is mounted and build a full
patch. In configurations with several roots this caused remove of the
sources from other roots.

Another problem was that a patchProcess closed a patch builder which
adds the EXIT command to the patch. When agent applies a patch it stops
once the first EXIT is encountered, as a result it applied sources only
from the first VCS root.