TW-38247 fix agent-side checkout when tags are used as branches
In case of tags git-plugin didn't updated working directory to the
revision in build. It worked most of the time because tags usually don't
change their revision. But it doesn't work in case of per-checking
triggering or when custom build is ran on a revision. Now we checkout
revision in build for tags explicitly if it is different from the tag's
revision.