Use simplified git-specific logger and make it mandatory
The real BuildProgressLogger is not always available, e.g. there is no
logger during git version detection on agent start. The simplified
logger is easier to mock, it can also hide real logger quirks like flows
vs blocks and the fact that empty blocks are not shown in the log.