TW-40421 share repositories between threads
Before that we used the jgit's RepositoryCache where a call to close()
removes a repository from the cache immediately, so it was possible that
different threads had different instances of the same repository wasting
memory. Switch to the cache tracking repository usages where a call to
release() removes repository only when no thread uses it.