TW-42105 labeling should not block changes collecting
For some reason a push of a single tag can take tens of minutes in a big
repository (TW-40639). We ran a push with write lock held, changes
collecting also uses this lock to ensure only one thread is writing into
repository.

Write lock doesn't seem to be necessary here: it protects the local
repository and we are pushing, ie. writing into the remote repository
not the local one.

Now labeling takes only the rmLock to prevent cleanup from removing
repository while labeling is in progress, it should not delay changes
collecting anymore.