TW-40639 speed-up labeling
Labeling is slow in big repositories, because in order to compute which
objects should be sent to remote repository we have to mark all objects
available in the remote repository as uninteresting.

When TeamCity labels commit which is already present in the remote
repository (the most common case for labeling) we only need to push the
tag object itself, we can do that without marking uninteresting objects.