Stop implementing old collect changes interface
When collect changes policy implements several interfaces TeamCity
prefers the interface which collects changes between repository states
over the interface collecting changes between single
revisions. Therefore CollectChangesBetweenRoots implementation can be
removed as it isn't used in runtime.

A few tests relied on the logic presented only in the old interface
implementation: when a fromVersion wasn't found in repository we
reported last 10 commits reachable from the toVersion. This logic
doesn't seem to be useful, especially in a multi-branch
scenario (hundred of branches will give us thousand of changes, some of
them can be old and so on), so we drop the logic. Tests are changed to
match the current logic which reports no changes in this case.