TW-43884 detect mirror remove failures
When repeating git mirror fetch make sure the old repository clone is
successfully removed. Without such a check we can get a situation when
some files are removed and some are not, this can cause the
'refs/heads/xxx does not point to a valid object!' error during
subsequent fetches. If remove failed - clone the repository in another
fresh directory.