TW-44944 handle 'git ls-remote' error
Don't suppress the error. If we do, then all refs will be considered as
outdated and will be removed. If there are thousands of refs this can be
slow, and after that they will be re-fetched. Instead of suppressing run
several attemps to list remote refs and fail checkout if they all fail.